Crate rusoto_workspaces

Amazon WorkSpaces Service
Amazon WorkSpaces enables you to provision virtual, cloud-based Microsoft Windows and Amazon Linux desktops for your users.

If you're using the service, you're probably looking for WorkspacesClient and Workspaces.

Structs

AccountModification
Describes a modification to the configuration of bring your own license (BYOL) for the specified account.

AssociateIpGroupsRequest

AssociateIpGroupsResult

AuthorizeIpRulesRequest

AuthorizeIpRulesResult

ClientProperties
Describes an Amazon WorkSpaces client.

ClientPropertiesResult
Information about the Amazon WorkSpaces client.

ComputeType
Describes the compute type.

CreateIpGroupRequest

CreateIpGroupResult

CreateTagsRequest

CreateTagsResult

CreateWorkspacesRequest

CreateWorkspacesResult

DefaultWorkspaceCreationProperties
Describes the default values used to create a WorkSpace.

DeleteIpGroupRequest

DeleteIpGroupResult

DeleteTagsRequest

DeleteTagsResult

DeleteWorkspaceImageRequest

DeleteWorkspaceImageResult

DescribeAccountModificationsRequest

DescribeAccountModificationsResult

DescribeAccountRequest

DescribeAccountResult

DescribeClientPropertiesRequest

DescribeClientPropertiesResult

DescribeIpGroupsRequest

DescribeIpGroupsResult

DescribeTagsRequest

DescribeTagsResult

DescribeWorkspaceBundlesRequest

DescribeWorkspaceBundlesResult

DescribeWorkspaceDirectoriesRequest

DescribeWorkspaceDirectoriesResult

DescribeWorkspaceImagesRequest

DescribeWorkspaceImagesResult

DescribeWorkspacesConnectionStatusRequest

DescribeWorkspacesConnectionStatusResult

DescribeWorkspacesRequest

DescribeWorkspacesResult

DisassociateIpGroupsRequest

DisassociateIpGroupsResult

FailedCreateWorkspaceRequest
Describes a WorkSpace that cannot be created.

FailedWorkspaceChangeRequest
Describes a WorkSpace that could not be rebooted. (RebootWorkspaces), rebuilt (RebuildWorkspaces), terminated (TerminateWorkspaces), started (StartWorkspaces), or stopped (StopWorkspaces).

ImportWorkspaceImageRequest

ImportWorkspaceImageResult

IpRuleItem
Describes a rule for an IP access control group.

ListAvailableManagementCidrRangesRequest

ListAvailableManagementCidrRangesResult

ModificationState
Describes a WorkSpace modification.

ModifyAccountRequest

ModifyAccountResult

ModifyClientPropertiesRequest

ModifyClientPropertiesResult

ModifyWorkspacePropertiesRequest

ModifyWorkspacePropertiesResult

ModifyWorkspaceStateRequest

ModifyWorkspaceStateResult

OperatingSystem
The operating system that the image is running.

RebootRequest
Describes the information used to reboot a WorkSpace.

RebootWorkspacesRequest

RebootWorkspacesResult

RebuildRequest
Describes the information used to rebuild a WorkSpace.

RebuildWorkspacesRequest

RebuildWorkspacesResult

RevokeIpRulesRequest

RevokeIpRulesResult

RootStorage
Describes the root volume for a WorkSpace bundle.

StartRequest
Information used to start a WorkSpace.

StartWorkspacesRequest

StartWorkspacesResult

StopRequest
Describes the information used to stop a WorkSpace.

StopWorkspacesRequest

StopWorkspacesResult

Tag
Describes a tag.

TerminateRequest
Describes the information used to terminate a WorkSpace.

TerminateWorkspacesRequest

TerminateWorkspacesResult

UpdateRulesOfIpGroupRequest

UpdateRulesOfIpGroupResult

UserStorage
Describes the user storage for a WorkSpace bundle.

Workspace
Describes a WorkSpace.

WorkspaceBundle
Describes a WorkSpace bundle.

WorkspaceConnectionStatus
Describes the connection status of a WorkSpace.

WorkspaceDirectory
Describes an AWS Directory Service directory that is used with Amazon WorkSpaces.

WorkspaceImage
Describes a WorkSpace image.

WorkspaceProperties
Describes a WorkSpace.

WorkspaceRequest
Describes the information used to create a WorkSpace.

WorkspacesClient
A client for the Amazon WorkSpaces API.

WorkspacesIpGroup
Describes an IP access control group.

Enums

AssociateIpGroupsError
Errors returned by AssociateIpGroups

AuthorizeIpRulesError
Errors returned by AuthorizeIpRules

CreateIpGroupError
Errors returned by CreateIpGroup

CreateTagsError
Errors returned by CreateTags

CreateWorkspacesError
Errors returned by CreateWorkspaces

DeleteIpGroupError
Errors returned by DeleteIpGroup

DeleteTagsError
Errors returned by DeleteTags

DeleteWorkspaceImageError
Errors returned by DeleteWorkspaceImage

DescribeAccountError
Errors returned by DescribeAccount

DescribeAccountModificationsError
Errors returned by DescribeAccountModifications

DescribeClientPropertiesError
Errors returned by DescribeClientProperties

DescribeIpGroupsError
Errors returned by DescribeIpGroups

DescribeTagsError
Errors returned by DescribeTags

DescribeWorkspaceBundlesError
Errors returned by DescribeWorkspaceBundles

DescribeWorkspaceDirectoriesError
Errors returned by DescribeWorkspaceDirectories

DescribeWorkspaceImagesError
Errors returned by DescribeWorkspaceImages

DescribeWorkspacesConnectionStatusError
Errors returned by DescribeWorkspacesConnectionStatus

DescribeWorkspacesError
Errors returned by DescribeWorkspaces

DisassociateIpGroupsError
Errors returned by DisassociateIpGroups

ImportWorkspaceImageError
Errors returned by ImportWorkspaceImage

ListAvailableManagementCidrRangesError
Errors returned by ListAvailableManagementCidrRanges

ModifyAccountError
Errors returned by ModifyAccount

ModifyClientPropertiesError
Errors returned by ModifyClientProperties

ModifyWorkspacePropertiesError
Errors returned by ModifyWorkspaceProperties

ModifyWorkspaceStateError
Errors returned by ModifyWorkspaceState

RebootWorkspacesError
Errors returned by RebootWorkspaces

RebuildWorkspacesError
Errors returned by RebuildWorkspaces

RevokeIpRulesError
Errors returned by RevokeIpRules

StartWorkspacesError
Errors returned by StartWorkspaces

StopWorkspacesError
Errors returned by StopWorkspaces

TerminateWorkspacesError
Errors returned by TerminateWorkspaces

UpdateRulesOfIpGroupError
Errors returned by UpdateRulesOfIpGroup

Traits

Workspaces
Trait representing the capabilities of the Amazon WorkSpaces API. Amazon WorkSpaces clients implement this trait.