Struct rusoto_workspaces::WorkspaceDirectory

pub struct WorkspaceDirectory {
    pub alias: Option<String>,
    pub customer_user_name: Option<String>,
    pub directory_id: Option<String>,
    pub directory_name: Option<String>,
    pub directory_type: Option<String>,
    pub dns_ip_addresses: Option<Vec<String>>,
    pub iam_role_id: Option<String>,
    pub registration_code: Option<String>,
    pub state: Option<String>,
    pub subnet_ids: Option<Vec<String>>,
    pub workspace_creation_properties: Option<DefaultWorkspaceCreationProperties>,
    pub workspace_security_group_id: Option<String>,
}

Contains information about an AWS Directory Service directory for use with Amazon WorkSpaces.

Fields
alias: Option<String>

The directory alias.

customer_user_name: Option<String>

The user name for the service account.

directory_id: Option<String>

The directory identifier.

directory_name: Option<String>

The name of the directory.

directory_type: Option<String>

The directory type.

dns_ip_addresses: Option<Vec<String>>

An array of strings that contains the IP addresses of the DNS servers for the directory.

iam_role_id: Option<String>

The identifier of the IAM role. This is the role that allows Amazon WorkSpaces to make calls to other services, such as Amazon EC2, on your behalf.

registration_code: Option<String>

The registration code for the directory. This is the code that users enter in their Amazon WorkSpaces client application to connect to the directory.

state: Option<String>

The state of the directory's registration with Amazon WorkSpaces

subnet_ids: Option<Vec<String>>

An array of strings that contains the identifiers of the subnets used with the directory.

workspace_creation_properties: Option<DefaultWorkspaceCreationProperties>

A structure that specifies the default creation properties for all WorkSpaces in the directory.

workspace_security_group_id: Option<String>

The identifier of the security group that is assigned to new WorkSpaces.