Struct rusoto_workspaces::WorkspacesClient

pub struct WorkspacesClient<P = CredentialsProvider, D = RequestDispatcher> where
    P: ProvideAwsCredentials,
    D: DispatchSignedRequest,  { /* fields omitted */ }

A client for the Amazon WorkSpaces API.

Methods

impl WorkspacesClient

pub fn simple(region: Region) -> WorkspacesClient

Creates a simple client backed by an implicit event loop.

The client will use the default credentials provider and tls client.

See the rusoto_core::reactor module for more details.

impl<P, D> WorkspacesClient<P, D> where
P: ProvideAwsCredentials,
D: DispatchSignedRequest,

pub fn new(
request_dispatcher: D,
credentials_provider: P,
region: Region
) -> Self