Trait rusoto_workspaces::Workspaces

pub trait Workspaces {
    fn create_tags(
        &self, 
        input: CreateTagsRequest
    ) -> RusotoFuture<CreateTagsResult, CreateTagsError>;

    fn create_workspaces(
        &self, 
        input: CreateWorkspacesRequest
    ) -> RusotoFuture<CreateWorkspacesResult, CreateWorkspacesError>;

    fn delete_tags(
        &self, 
        input: DeleteTagsRequest
    ) -> RusotoFuture<DeleteTagsResult, DeleteTagsError>;

    fn describe_tags(
        &self, 
        input: DescribeTagsRequest
    ) -> RusotoFuture<DescribeTagsResult, DescribeTagsError>;

    fn describe_workspace_bundles(
        &self, 
        input: DescribeWorkspaceBundlesRequest
    ) -> RusotoFuture<DescribeWorkspaceBundlesResult, DescribeWorkspaceBundlesError>;

    fn describe_workspace_directories(
        &self, 
        input: DescribeWorkspaceDirectoriesRequest
    ) -> RusotoFuture<DescribeWorkspaceDirectoriesResult, DescribeWorkspaceDirectoriesError>;

    fn describe_workspaces(
        &self, 
        input: DescribeWorkspacesRequest
    ) -> RusotoFuture<DescribeWorkspacesResult, DescribeWorkspacesError>;

    fn describe_workspaces_connection_status(
        &self, 
        input: DescribeWorkspacesConnectionStatusRequest
    ) -> RusotoFuture<DescribeWorkspacesConnectionStatusResult, DescribeWorkspacesConnectionStatusError>;

    fn modify_workspace_properties(
        &self, 
        input: ModifyWorkspacePropertiesRequest
    ) -> RusotoFuture<ModifyWorkspacePropertiesResult, ModifyWorkspacePropertiesError>;

    fn reboot_workspaces(
        &self, 
        input: RebootWorkspacesRequest
    ) -> RusotoFuture<RebootWorkspacesResult, RebootWorkspacesError>;

    fn rebuild_workspaces(
        &self, 
        input: RebuildWorkspacesRequest
    ) -> RusotoFuture<RebuildWorkspacesResult, RebuildWorkspacesError>;

    fn start_workspaces(
        &self, 
        input: StartWorkspacesRequest
    ) -> RusotoFuture<StartWorkspacesResult, StartWorkspacesError>;

    fn stop_workspaces(
        &self, 
        input: StopWorkspacesRequest
    ) -> RusotoFuture<StopWorkspacesResult, StopWorkspacesError>;

    fn terminate_workspaces(
        &self, 
        input: TerminateWorkspacesRequest
    ) -> RusotoFuture<TerminateWorkspacesResult, TerminateWorkspacesError>;
}

Trait representing the capabilities of the Amazon WorkSpaces API. Amazon WorkSpaces clients implement this trait.

Required Methods

fn create_tags(
&self,
input: CreateTagsRequest
) -> RusotoFuture<CreateTagsResult, CreateTagsError>

Creates tags for the specified WorkSpace.

fn create_workspaces(
&self,
input: CreateWorkspacesRequest
) -> RusotoFuture<CreateWorkspacesResult, CreateWorkspacesError>

Creates one or more WorkSpaces.

This operation is asynchronous and returns before the WorkSpaces are created.

fn delete_tags(
&self,
input: DeleteTagsRequest
) -> RusotoFuture<DeleteTagsResult, DeleteTagsError>

Deletes the specified tags from a WorkSpace.

fn describe_tags(
&self,
input: DescribeTagsRequest
) -> RusotoFuture<DescribeTagsResult, DescribeTagsError>

Describes the tags for the specified WorkSpace.

fn describe_workspace_bundles(
&self,
input: DescribeWorkspaceBundlesRequest
) -> RusotoFuture<DescribeWorkspaceBundlesResult, DescribeWorkspaceBundlesError>

Describes the available WorkSpace bundles.

You can filter the results using either bundle ID or owner, but not both.

fn describe_workspace_directories(
&self,
input: DescribeWorkspaceDirectoriesRequest
) -> RusotoFuture<DescribeWorkspaceDirectoriesResult, DescribeWorkspaceDirectoriesError>

Describes the available AWS Directory Service directories that are registered with Amazon WorkSpaces.

fn describe_workspaces(
&self,
input: DescribeWorkspacesRequest
) -> RusotoFuture<DescribeWorkspacesResult, DescribeWorkspacesError>

Describes the specified WorkSpaces.

You can filter the results using bundle ID, directory ID, or owner, but you can specify only one filter at a time.

fn describe_workspaces_connection_status(
&self,
input: DescribeWorkspacesConnectionStatusRequest
) -> RusotoFuture<DescribeWorkspacesConnectionStatusResult, DescribeWorkspacesConnectionStatusError>

Describes the connection status of the specified WorkSpaces.

fn modify_workspace_properties(
&self,
input: ModifyWorkspacePropertiesRequest
) -> RusotoFuture<ModifyWorkspacePropertiesResult, ModifyWorkspacePropertiesError>

Modifies the specified WorkSpace properties.

fn reboot_workspaces(
&self,
input: RebootWorkspacesRequest
) -> RusotoFuture<RebootWorkspacesResult, RebootWorkspacesError>

Reboots the specified WorkSpaces.

You cannot reboot a WorkSpace unless its state is AVAILABLE, IMPAIRED, or INOPERABLE.

This operation is asynchronous and returns before the WorkSpaces have rebooted.

fn rebuild_workspaces(
&self,
input: RebuildWorkspacesRequest
) -> RusotoFuture<RebuildWorkspacesResult, RebuildWorkspacesError>

Rebuilds the specified WorkSpaces.

You cannot rebuild a WorkSpace unless its state is AVAILABLE or ERROR.

Rebuilding a WorkSpace is a potentially destructive action that can result in the loss of data. For more information, see Rebuild a WorkSpace.

This operation is asynchronous and returns before the WorkSpaces have been completely rebuilt.

fn start_workspaces(
&self,
input: StartWorkspacesRequest
) -> RusotoFuture<StartWorkspacesResult, StartWorkspacesError>

Starts the specified WorkSpaces.

You cannot start a WorkSpace unless it has a running mode of AutoStop and a state of STOPPED.

fn stop_workspaces(
&self,
input: StopWorkspacesRequest
) -> RusotoFuture<StopWorkspacesResult, StopWorkspacesError>

Stops the specified WorkSpaces.

You cannot stop a WorkSpace unless it has a running mode of AutoStop and a state of AVAILABLE, IMPAIRED, UNHEALTHY, or ERROR.

fn terminate_workspaces(
&self,
input: TerminateWorkspacesRequest
) -> RusotoFuture<TerminateWorkspacesResult, TerminateWorkspacesError>

Terminates the specified WorkSpaces.

Terminating a WorkSpace is a permanent action and cannot be undone. The user's data is destroyed. If you need to archive any user data, contact Amazon Web Services before terminating the WorkSpace.

You can terminate a WorkSpace that is in any state except SUSPENDED.

This operation is asynchronous and returns before the WorkSpaces have been completely terminated.