Struct rusoto_workspaces::WorkspacesClient

pub struct WorkspacesClient { /* fields omitted */ }

A client for the Amazon WorkSpaces API.

Methods

impl WorkspacesClient

pub fn new(region: Region) -> WorkspacesClient

Creates a client backed by the default tokio event loop.

The client will use the default credentials provider and tls client.

pub fn new_with<P, D>(
request_dispatcher: D,
credentials_provider: P,
region: Region
) -> WorkspacesClient where
P: ProvideAwsCredentials + Send + Sync + 'static,
P::Future: Send,
D: DispatchSignedRequest + Send + Sync + 'static,
D::Future: Send,