Trait rusoto_workspaces::Workspaces

pub trait Workspaces {
    fn associate_ip_groups(
        &self, 
        input: AssociateIpGroupsRequest
    ) -> RusotoFuture<AssociateIpGroupsResult, AssociateIpGroupsError>;

    fn authorize_ip_rules(
        &self, 
        input: AuthorizeIpRulesRequest
    ) -> RusotoFuture<AuthorizeIpRulesResult, AuthorizeIpRulesError>;

    fn create_ip_group(
        &self, 
        input: CreateIpGroupRequest
    ) -> RusotoFuture<CreateIpGroupResult, CreateIpGroupError>;

    fn create_tags(
        &self, 
        input: CreateTagsRequest
    ) -> RusotoFuture<CreateTagsResult, CreateTagsError>;

    fn create_workspaces(
        &self, 
        input: CreateWorkspacesRequest
    ) -> RusotoFuture<CreateWorkspacesResult, CreateWorkspacesError>;

    fn delete_ip_group(
        &self, 
        input: DeleteIpGroupRequest
    ) -> RusotoFuture<DeleteIpGroupResult, DeleteIpGroupError>;

    fn delete_tags(
        &self, 
        input: DeleteTagsRequest
    ) -> RusotoFuture<DeleteTagsResult, DeleteTagsError>;

    fn describe_ip_groups(
        &self, 
        input: DescribeIpGroupsRequest
    ) -> RusotoFuture<DescribeIpGroupsResult, DescribeIpGroupsError>;

    fn describe_tags(
        &self, 
        input: DescribeTagsRequest
    ) -> RusotoFuture<DescribeTagsResult, DescribeTagsError>;

    fn describe_workspace_bundles(
        &self, 
        input: DescribeWorkspaceBundlesRequest
    ) -> RusotoFuture<DescribeWorkspaceBundlesResult, DescribeWorkspaceBundlesError>;

    fn describe_workspace_directories(
        &self, 
        input: DescribeWorkspaceDirectoriesRequest
    ) -> RusotoFuture<DescribeWorkspaceDirectoriesResult, DescribeWorkspaceDirectoriesError>;

    fn describe_workspaces(
        &self, 
        input: DescribeWorkspacesRequest
    ) -> RusotoFuture<DescribeWorkspacesResult, DescribeWorkspacesError>;

    fn describe_workspaces_connection_status(
        &self, 
        input: DescribeWorkspacesConnectionStatusRequest
    ) -> RusotoFuture<DescribeWorkspacesConnectionStatusResult, DescribeWorkspacesConnectionStatusError>;

    fn disassociate_ip_groups(
        &self, 
        input: DisassociateIpGroupsRequest
    ) -> RusotoFuture<DisassociateIpGroupsResult, DisassociateIpGroupsError>;

    fn modify_workspace_properties(
        &self, 
        input: ModifyWorkspacePropertiesRequest
    ) -> RusotoFuture<ModifyWorkspacePropertiesResult, ModifyWorkspacePropertiesError>;

    fn modify_workspace_state(
        &self, 
        input: ModifyWorkspaceStateRequest
    ) -> RusotoFuture<ModifyWorkspaceStateResult, ModifyWorkspaceStateError>;

    fn reboot_workspaces(
        &self, 
        input: RebootWorkspacesRequest
    ) -> RusotoFuture<RebootWorkspacesResult, RebootWorkspacesError>;

    fn rebuild_workspaces(
        &self, 
        input: RebuildWorkspacesRequest
    ) -> RusotoFuture<RebuildWorkspacesResult, RebuildWorkspacesError>;

    fn revoke_ip_rules(
        &self, 
        input: RevokeIpRulesRequest
    ) -> RusotoFuture<RevokeIpRulesResult, RevokeIpRulesError>;

    fn start_workspaces(
        &self, 
        input: StartWorkspacesRequest
    ) -> RusotoFuture<StartWorkspacesResult, StartWorkspacesError>;

    fn stop_workspaces(
        &self, 
        input: StopWorkspacesRequest
    ) -> RusotoFuture<StopWorkspacesResult, StopWorkspacesError>;

    fn terminate_workspaces(
        &self, 
        input: TerminateWorkspacesRequest
    ) -> RusotoFuture<TerminateWorkspacesResult, TerminateWorkspacesError>;

    fn update_rules_of_ip_group(
        &self, 
        input: UpdateRulesOfIpGroupRequest
    ) -> RusotoFuture<UpdateRulesOfIpGroupResult, UpdateRulesOfIpGroupError>;
}

Trait representing the capabilities of the Amazon WorkSpaces API. Amazon WorkSpaces clients implement this trait.

Required Methods

fn associate_ip_groups(
&self,
input: AssociateIpGroupsRequest
) -> RusotoFuture<AssociateIpGroupsResult, AssociateIpGroupsError>

Associates the specified IP access control group with the specified directory.

fn authorize_ip_rules(
&self,
input: AuthorizeIpRulesRequest
) -> RusotoFuture<AuthorizeIpRulesResult, AuthorizeIpRulesError>

Adds one or more rules to the specified IP access control group.

This action gives users permission to access their WorkSpaces from the CIDR address ranges specified in the rules.

fn create_ip_group(
&self,
input: CreateIpGroupRequest
) -> RusotoFuture<CreateIpGroupResult, CreateIpGroupError>

Creates an IP access control group.

An IP access control group provides you with the ability to control the IP addresses from which users are allowed to access their WorkSpaces. To specify the CIDR address ranges, add rules to your IP access control group and then associate the group with your directory. You can add rules when you create the group or at any time using AuthorizeIpRules.

There is a default IP access control group associated with your directory. If you don't associate an IP access control group with your directory, the default group is used. The default group includes a default rule that allows users to access their WorkSpaces from anywhere. You cannot modify the default IP access control group for your directory.

fn create_tags(
&self,
input: CreateTagsRequest
) -> RusotoFuture<CreateTagsResult, CreateTagsError>

Creates the specified tags for the specified WorkSpace.

fn create_workspaces(
&self,
input: CreateWorkspacesRequest
) -> RusotoFuture<CreateWorkspacesResult, CreateWorkspacesError>

Creates one or more WorkSpaces.

This operation is asynchronous and returns before the WorkSpaces are created.

fn delete_ip_group(
&self,
input: DeleteIpGroupRequest
) -> RusotoFuture<DeleteIpGroupResult, DeleteIpGroupError>

Deletes the specified IP access control group.

You cannot delete an IP access control group that is associated with a directory.

fn delete_tags(
&self,
input: DeleteTagsRequest
) -> RusotoFuture<DeleteTagsResult, DeleteTagsError>

Deletes the specified tags from the specified WorkSpace.

fn describe_ip_groups(
&self,
input: DescribeIpGroupsRequest
) -> RusotoFuture<DescribeIpGroupsResult, DescribeIpGroupsError>

Describes one or more of your IP access control groups.

fn describe_tags(
&self,
input: DescribeTagsRequest
) -> RusotoFuture<DescribeTagsResult, DescribeTagsError>

Describes the specified tags for the specified WorkSpace.

fn describe_workspace_bundles(
&self,
input: DescribeWorkspaceBundlesRequest
) -> RusotoFuture<DescribeWorkspaceBundlesResult, DescribeWorkspaceBundlesError>

Describes the available WorkSpace bundles.

You can filter the results using either bundle ID or owner, but not both.

fn describe_workspace_directories(
&self,
input: DescribeWorkspaceDirectoriesRequest
) -> RusotoFuture<DescribeWorkspaceDirectoriesResult, DescribeWorkspaceDirectoriesError>

Describes the available AWS Directory Service directories that are registered with Amazon WorkSpaces.

fn describe_workspaces(
&self,
input: DescribeWorkspacesRequest
) -> RusotoFuture<DescribeWorkspacesResult, DescribeWorkspacesError>

Describes the specified WorkSpaces.

You can filter the results using bundle ID, directory ID, or owner, but you can specify only one filter at a time.

fn describe_workspaces_connection_status(
&self,
input: DescribeWorkspacesConnectionStatusRequest
) -> RusotoFuture<DescribeWorkspacesConnectionStatusResult, DescribeWorkspacesConnectionStatusError>

Describes the connection status of the specified WorkSpaces.

fn disassociate_ip_groups(
&self,
input: DisassociateIpGroupsRequest
) -> RusotoFuture<DisassociateIpGroupsResult, DisassociateIpGroupsError>

Disassociates the specified IP access control group from the specified directory.

fn modify_workspace_properties(
&self,
input: ModifyWorkspacePropertiesRequest
) -> RusotoFuture<ModifyWorkspacePropertiesResult, ModifyWorkspacePropertiesError>

Modifies the specified WorkSpace properties.

fn modify_workspace_state(
&self,
input: ModifyWorkspaceStateRequest
) -> RusotoFuture<ModifyWorkspaceStateResult, ModifyWorkspaceStateError>

Sets the state of the specified WorkSpace.

To maintain a WorkSpace without being interrupted, set the WorkSpace state to ADMIN_MAINTENANCE. WorkSpaces in this state do not respond to requests to reboot, stop, start, or rebuild. An AutoStop WorkSpace in this state is not stopped. Users can log into a WorkSpace in the ADMIN_MAINTENANCE state.

fn reboot_workspaces(
&self,
input: RebootWorkspacesRequest
) -> RusotoFuture<RebootWorkspacesResult, RebootWorkspacesError>

Reboots the specified WorkSpaces.

You cannot reboot a WorkSpace unless its state is AVAILABLE or UNHEALTHY.

This operation is asynchronous and returns before the WorkSpaces have rebooted.

fn rebuild_workspaces(
&self,
input: RebuildWorkspacesRequest
) -> RusotoFuture<RebuildWorkspacesResult, RebuildWorkspacesError>

Rebuilds the specified WorkSpace.

You cannot rebuild a WorkSpace unless its state is AVAILABLE, ERROR, or UNHEALTHY.

Rebuilding a WorkSpace is a potentially destructive action that can result in the loss of data. For more information, see Rebuild a WorkSpace.

This operation is asynchronous and returns before the WorkSpaces have been completely rebuilt.

fn revoke_ip_rules(
&self,
input: RevokeIpRulesRequest
) -> RusotoFuture<RevokeIpRulesResult, RevokeIpRulesError>

Removes one or more rules from the specified IP access control group.

fn start_workspaces(
&self,
input: StartWorkspacesRequest
) -> RusotoFuture<StartWorkspacesResult, StartWorkspacesError>

Starts the specified WorkSpaces.

You cannot start a WorkSpace unless it has a running mode of AutoStop and a state of STOPPED.

fn stop_workspaces(
&self,
input: StopWorkspacesRequest
) -> RusotoFuture<StopWorkspacesResult, StopWorkspacesError>

Stops the specified WorkSpaces.

You cannot stop a WorkSpace unless it has a running mode of AutoStop and a state of AVAILABLE, IMPAIRED, UNHEALTHY, or ERROR.

fn terminate_workspaces(
&self,
input: TerminateWorkspacesRequest
) -> RusotoFuture<TerminateWorkspacesResult, TerminateWorkspacesError>

Terminates the specified WorkSpaces.

Terminating a WorkSpace is a permanent action and cannot be undone. The user's data is destroyed. If you need to archive any user data, contact Amazon Web Services before terminating the WorkSpace.

You can terminate a WorkSpace that is in any state except SUSPENDED.

This operation is asynchronous and returns before the WorkSpaces have been completely terminated.

fn update_rules_of_ip_group(
&self,
input: UpdateRulesOfIpGroupRequest
) -> RusotoFuture<UpdateRulesOfIpGroupResult, UpdateRulesOfIpGroupError>

Replaces the current rules of the specified IP access control group with the specified rules.